+++ CHANGELOG.md
@@ -24,6 +24,8 @@
 * Add `refund_from_transaction_fee_currency_iso_code` field to paypal_details
 * Add `revoked_at` field to paypal_account
 * Add support for `PaymentMethodRevokedByCustomer` webhook
+* Add `payment_method_nonce` field to `LocalPaymentCompleted` webhook
+* Add `transaction` field to `LocalPaymentCompleted` webhook
 
 # 2.95.0
 * Deprecate `GrantedPaymentInstrumentUpdate` and add `GrantorUpdatedGrantedPaymentMethod` and `RecipientUpdatedGrantedPaymentMethod`